model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
pymongo==4.6.0
sentence-transformers>=2.7.0
scikit-learn==1.5.1
numpy>=1.26.0
python-dotenv==1.0.0
tf-keras>=2.20.0
